hotel/requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
cssselect2==0.7.0
distlib==0.3.6
Django==4.1.5
django-bootstrap-datepicker-plus==5.0.2
django-bootstrap-v5==1.0.11
django-mathfilters==1.0.0
django-payments==1.0.0
django-phonenumber-field==7.0.1
django-shortcuts==1.6
django-widget-tweaks==1.4.8
django_debug_toolbar==3.8.1
filelock==3.8.0
future==0.18.2
html5lib==1.1
idna==3.4
lxml==4.9.2
oscrypto==1.3.0
package-name==0.1
phonenumberslite==8.13.4
Pillow==9.3.0
platformdirs==2.5.4
pycparser==2.21
pydantic==1.10.2
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.13
virtualenv==20.16.7
webencodings==0.5.1
xhtml2pdf==0.2.8
